+++ src/OFS/Lockable.py
@@ -13,6 +13,7 @@
 
 from AccessControl.class_init import InitializeClass
 from AccessControl.Permissions import webdav_lock_items
+from AccessControl.Permissions import webdav_manage_locks
 from AccessControl.Permissions import webdav_unlock_items
 from AccessControl.SecurityInfo import ClassSecurityInfo
 from Acquisition import aq_base
@@ -20,7 +21,6 @@ from OFS.EtagSupport import EtagSupport
 from OFS.interfaces import ILockItem
 from OFS.interfaces import IWriteLock
 from Persistence import PersistentMapping
-from webdav import webdav_manage_locks
 from zope.interface import implementer